Re #26960 Convert MoveSANSInst. class to free function
Converts the MoveSANSInst. class to a free function, as the state we
need does not need to be stored as callers always have access to it
