+++ Framework/PythonInterface/plugins/algorithms/WorkflowAlgorithms/SANS/SANSBeamCentreFinderCore.py
@@ -12,7 +12,7 @@ from __future__ import (absolute_import, division, print_function)
 from mantid.api import (DataProcessorAlgorithm, MatrixWorkspaceProperty, AlgorithmFactory, PropertyMode, Progress,
                         IEventWorkspace)
 from mantid.kernel import (Direction, PropertyManagerProperty, StringListValidator)
-from sans.algorithm_detail.MoveSansInstrumentComponent import MoveSansInstrumentComponent, MoveTypes
+from sans.algorithm_detail.move_sans_instrument_component import move_component, MoveTypes
 from sans.common.constants import EMPTY_NAME
 from sans.common.general_functions import create_child_algorithm, append_to_sans_file_tag
 from sans.state.state_base import create_deserialized_sans_state_from_property_manager
@@ -76,8 +76,8 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         self.declareProperty("Centre1", 0.0, direction=Direction.Input)
         self.declareProperty("Centre2", 0.0, direction=Direction.Input)
 
-        self.declareProperty("RMax", 0.26, direction= Direction.Input)
-        self.declareProperty("RMin", 0.06, direction= Direction.Input)
+        self.declareProperty("RMax", 0.26, direction=Direction.Input)
+        self.declareProperty("RMin", 0.06, direction=Direction.Input)
 
         # ----------
         # OUTPUT
@@ -113,7 +113,7 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         state.mask.use_mask_phi_mirror = True
 
         # Set compatibility mode
-        #state.compatibility.use_compatibility_mode = self.getProperty('CompatibilityMode').value
+        # state.compatibility.use_compatibility_mode = self.getProperty('CompatibilityMode').value
 
         # Set test centre
         state.move.detectors[DetectorType.to_string(DetectorType.LAB)].sample_centre_pos1 = \
@@ -296,13 +296,10 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
     def _move(self, state, workspace, component, is_transmission=False):
         # First we set the workspace to zero, since it might have been moved around by the user in the ADS
         # Second we use the initial move to bring the workspace into the correct position
-        reset_alg = MoveSansInstrumentComponent(move_type=MoveTypes.RESET_POSITION)
-        reset_alg.move_component(move_info=state.move, component_name='',
-                                 workspace=workspace)
-
-        move_alg = MoveSansInstrumentComponent(move_type=MoveTypes.INITIAL_MOVE)
-        move_alg.move_component(component_name=component, move_info=state.move,
-                                workspace=workspace, is_transmission_workspace=is_transmission)
+        move_component(move_info=state.move, component_name='', move_type=MoveTypes.RESET_POSITION,
+                       workspace=workspace)
+        move_component(component_name=component, move_info=state.move, move_type=MoveTypes.INITIAL_MOVE,
+                       workspace=workspace, is_transmission_workspace=is_transmission)
         return workspace
 
     def _mask(self, state_serialized, workspace, component):
@@ -364,7 +361,7 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         wavelength_adjustment = adjustment_alg.getProperty("OutputWorkspaceWavelengthAdjustment").value
         pixel_adjustment = adjustment_alg.getProperty("OutputWorkspacePixelAdjustment").value
         wavelength_and_pixel_adjustment = adjustment_alg.getProperty(
-                                           "OutputWorkspaceWavelengthAndPixelAdjustment").value
+            "OutputWorkspaceWavelengthAndPixelAdjustment").value
         return wavelength_adjustment, pixel_adjustment, wavelength_and_pixel_adjustment
 
     def _convert_to_histogram(self, workspace):
